requirements.1.txt
cffi==1.11.4
click==6.7
Flask==0.12.2
Flask-Misaka==0.4.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
misaka==2.1.0
mysql-connector-python==8.0.6
pandas==0.22.0
passlib==1.7.1
pycparser==2.18
python-dateutil==2.6.1
pytz==2018.3
scikit-learn==0.19.1
six==1.11.0
sklearn==0.0
uwsgi==2.0.17
Werkzeug==0.14.1
